requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
python-multipart
python-jose
bcrypt
email-validator
passlib
passlib[argon2]
PyJWT
pymysql
